refactor(cli): flatten banking command module
Move the Banking command slice to the neutral command root and update its internal export, test-only prefixes, and approved help copy without changing Banking API compatibility behavior.
